Functionality/user routes

+++ controllers/userController.js
@@ -16,13 +16,10 @@ module.exports = {
             .populate({path: 'thoughts', select: '-__v'})
             .populate({path: 'friends', select: '-__v'});
 
-            // res.json(user);
-
             if (!user) {
                 return res.status(404).json({ message: 'No post with that ID' });
-              }
-        
-            res.json(user);
+              }      
+            res.status(200).json(user);
         } catch(err) {
             console.log(err);
             res.status(500).json(err);
@@ -31,15 +28,23 @@ module.exports = {
     async createUser(req,res) {
         try{
             const createUser = await User.create(req.body);
-            res.json(createUser);
+            res.status(200).json(createUser);
         } catch(err) {
             res.status(500).json(err);
         }
     },
     async deleteUser(req,res) {
         try{
+            const userfinder = await User.findById(req.params.userId)
+
+            if (!userfinder){
+                return res.status(400).json({message: "No user with that Id exists"});
+            }
+            const thoughtFinder = await Thought.find({username: userfinder.username})
             const user = await User.deleteOne({_id: req.params.userId})
-            res.json(user);
+            const thought = await Thought.deleteMany({username: userfinder.username})
+            res.status(200).json({Message: "User and associated thoughts deleted"});
+
         } catch(err){
             res.status(500).json(err);
         }
@@ -47,6 +52,11 @@ module.exports = {
     async updateUser(req,res) {
         try{
             const updUser = await User.findOneAndUpdate({_id: req.params.userId},req.body,{ new: true })
+
+            if(!updUser){
+                return res.status(400).json({message: "No user with this ID exists"})
+            }
+
             res.json(updUser);
         } catch(err){
             res.status(500).json(err);
